fix: InitialProjectFileGenerator
