+++ packages/create-amplify/src/package-manager-controller/npm_package_manager_controller.ts
@@ -1,10 +1,7 @@
 import { execa as _execa } from 'execa';
 import { executeWithDebugLogger } from '../execute_with_logger.js';
-import {
-  DependencyType,
-  PackageManagerController,
-  PackageManagerControllerFactory,
-} from './package_manager_controller_factory.js';
+import { DependencyType } from './package_manager_controller_factory.js';
+import { PackageManagerController } from './package_manager_controller.js';
 
 /**
  *
@@ -13,8 +10,8 @@ export class NpmPackageManagerController extends PackageManagerController {
   /**
    * Abstraction around npm commands that are needed to initialize a project and install dependencies
    */
-  constructor() {
-    super();
+  constructor(readonly projectRoot: string) {
+    super(projectRoot);
     this.executable = 'npm';
     this.binaryRunner = 'npx';
     this.installCommand = 'install';
