+++ packages/create-amplify/src/package-manager-controller/package_manager_controller_factory.test.ts
@@ -5,7 +5,7 @@ import assert from 'assert';
 void describe('PackageManagerController', () => {
   void it('executes expected dev dependency install command', async () => {
     const execaMock = mock.fn();
-    const npmPackageManagerController = new NpmPackageManagerController();
+    const npmPackageManagerController = new NpmPackageManagerController('./');
     await npmPackageManagerController.installDependencies(['testDep'], 'dev');
     assert.deepStrictEqual(execaMock.mock.calls[0].arguments, [
       'npm',
@@ -16,7 +16,7 @@ void describe('PackageManagerController', () => {
 
   void it('executes expected prod dependency install command', async () => {
     const execaMock = mock.fn();
-    const npmPackageManagerController = new NpmPackageManagerController();
+    const npmPackageManagerController = new NpmPackageManagerController('./');
     await npmPackageManagerController.installDependencies(['testDep'], 'prod');
     assert.deepStrictEqual(execaMock.mock.calls[0].arguments, [
       'npm',
@@ -26,7 +26,7 @@ void describe('PackageManagerController', () => {
   });
 
   void it('throws when installing dependencies rejects', async () => {
-    const npmPackageManagerController = new NpmPackageManagerController();
+    const npmPackageManagerController = new NpmPackageManagerController('./');
     await assert.rejects(
       () =>
         npmPackageManagerController.installDependencies(['testDep'], 'prod'),
@@ -37,7 +37,7 @@ void describe('PackageManagerController', () => {
   });
 
   void it('throws when installing dev dependencies rejects', async () => {
-    const npmPackageManagerController = new NpmPackageManagerController();
+    const npmPackageManagerController = new NpmPackageManagerController('./');
     await assert.rejects(
       () =>
         npmPackageManagerController.installDependencies(['testDevDep'], 'dev'),
