+++ packages/create-amplify/src/package-manager-controller/yarn_classic_package_manager_controller.ts
@@ -1,11 +1,8 @@
 import { existsSync as _existsSync } from 'fs';
 import { execa as _execa } from 'execa';
 import { executeWithDebugLogger } from '../execute_with_logger.js';
-import {
-  DependencyType,
-  PackageManagerController,
-  PackageManagerControllerFactory,
-} from './package_manager_controller_factory.js';
+import { DependencyType } from './package_manager_controller_factory.js';
+import { PackageManagerController } from './package_manager_controller.js';
 
 /**
  *
@@ -14,8 +11,8 @@ export class YarnClassicPackageManagerController extends PackageManagerControlle
   /**
    * Abstraction around yarn classic commands that are needed to initialize a project and install dependencies
    */
-  constructor() {
-    super();
+  constructor(readonly projectRoot: string) {
+    super(projectRoot);
     this.executable = 'yarn';
     this.binaryRunner = 'yarn';
     this.installCommand = 'add';
