+++ packages/create-amplify/src/package-manager-controller/yarn_modern_package_manager_controller.ts
@@ -1,12 +1,7 @@
 import { execa as _execa } from 'execa';
-import fsp from 'fs/promises';
-import * as path from 'path';
 import { executeWithDebugLogger } from '../execute_with_logger.js';
-import {
-  DependencyType,
-  PackageManagerController,
-  PackageManagerControllerFactory,
-} from './package_manager_controller_factory.js';
+import { DependencyType } from './package_manager_controller_factory.js';
+import { PackageManagerController } from './package_manager_controller.js';
 
 /**
  *
@@ -15,8 +10,8 @@ export class YarnModernPackageManagerController extends PackageManagerController
   /**
    * Abstraction around yarn modern commands that are needed to initialize a project and install dependencies
    */
-  constructor() {
-    super();
+  constructor(readonly projectRoot: string) {
+    super(projectRoot);
     this.executable = 'yarn';
     this.binaryRunner = 'yarn';
     this.installCommand = 'add';
